:construction: Partially working rebuild of travel request edit as page instead of as modal.
